Moved hcod solver outside pyopensot.cpp file in order to make bindings compile when the flag OPENSOT_SOTH_FRONT_END is False. Tested hcod solver in panda_ik.example. HCOD bindings are now in pyopnesot_hcod python lib.
