+++ bindings/python/pyopensot_hcod.cpp
@@ -1,0 +1,5 @@
+#include "solver_hcod.hpp"
+
+PYBIND11_MODULE(pyopensot_hcod, m) {
+    pyHCOD(m);
+}
