+++ bindings/python/solver.hpp
@@ -5,7 +5,6 @@
 #include <OpenSoT/solvers/eHQP.h>
 #include <OpenSoT/solvers/iHQP.h>
 #include <OpenSoT/solvers/nHQP.h>
-#include <OpenSoT/solvers/HCOD.h>
 #include <OpenSoT/solvers/BackEndFactory.h>
 
 namespace py = pybind11;
@@ -102,13 +101,3 @@ void pynHQP(py::module& m) {
        .def("setPerformSelectiveNullSpaceRegularization", py::overload_cast<bool>(&solvers::nHQP::setPerformSelectiveNullSpaceRegularization));
 }
 
-void pyHCOD(py::module& m) {
-    py::class_<solvers::HCOD, std::shared_ptr<solvers::HCOD>, OpenSoT::Solver<Eigen::MatrixXd, Eigen::VectorXd>>(m, "HCOD")
-        .def(py::init<OpenSoT::AutoStack&, const double>())
-        .def(py::init<OpenSoT::Solver<Eigen::MatrixXd, Eigen::VectorXd>::Stack&, OpenSoT::Solver<Eigen::MatrixXd, Eigen::VectorXd>::ConstraintPtr, const double>())
-        .def("solve", solve<Eigen::MatrixXd, Eigen::VectorXd>)
-        .def("setDisableWeightsComputation", &solvers::HCOD::setDisableWeightsComputation)
-        .def("getDisableWeightsComputation", &solvers::HCOD::getDisableWeightsComputation)
-        .def("setDamping", &solvers::HCOD::setDamping)
-        .def("printSOT", &solvers::HCOD::printSOT);
-}
